+++ src/main/java/com/appdynamics/extensions/webspheremq/metricscollector/QueueMetricsCollector.java
@@ -1,131 +1,108 @@
 package com.appdynamics.extensions.webspheremq.metricscollector;
 
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
+import com.appdynamics.extensions.conf.MonitorConfiguration;
 import com.appdynamics.extensions.util.metrics.MetricOverride;
-import com.appdynamics.extensions.webspheremq.config.QueueExcludeFilters;
-import com.appdynamics.extensions.webspheremq.config.QueueIncludeFilters;
+import com.appdynamics.extensions.webspheremq.config.ExcludeFilters;
 import com.appdynamics.extensions.webspheremq.config.QueueManager;
 import com.appdynamics.extensions.webspheremq.config.WMQMetricOverride;
-import com.ibm.mq.MQException;
 import com.ibm.mq.constants.CMQC;
 import com.ibm.mq.constants.CMQCFC;
-import com.ibm.mq.constants.MQConstants;
 import com.ibm.mq.pcf.PCFException;
 import com.ibm.mq.pcf.PCFMessage;
 import com.ibm.mq.pcf.PCFMessageAgent;
 import com.singularity.ee.agent.systemagent.api.AManagedMonitor;
 import com.singularity.ee.agent.systemagent.api.exception.TaskExecutionException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.Arrays;
+import java.util.Iterator;
+import java.util.Map;
+import java.util.Set;
 
 public class QueueMetricsCollector extends MetricsCollector {
 
 	public static final Logger logger = LoggerFactory.getLogger(QueueMetricsCollector.class);
 	private final String artifact = "Queues";
-	List<String> queueList;
 
-	public QueueMetricsCollector(Map<String, ? extends MetricOverride> metricsToReport, AManagedMonitor monitor, PCFMessageAgent agent, QueueManager queueManager, String metricPrefix) {
+	public QueueMetricsCollector(Map<String, ? extends MetricOverride> metricsToReport, MonitorConfiguration writer, PCFMessageAgent agent, QueueManager queueManager, String metricPrefix) {
 		this.metricsToReport = metricsToReport;
-		this.monitor = monitor;
+		this.writer = writer;
 		this.agent = agent;
 		this.metricPrefix = metricPrefix;
 		this.queueManager = queueManager;
 	}
 
 	@Override
-	protected void processFilter() throws TaskExecutionException {
-		List<String> allQueues = getQueueList();
-		logger.debug("All Queues: " + Arrays.toString(allQueues.toArray()));
-		// include all queues first in case include filter is not there.
-		List<String> includedQueues = allQueues;
-
-		// First evaluate include filters and then exclude filters
-		QueueIncludeFilters includeFilters = this.queueManager.getQueueIncludeFilters();
-		if (includeFilters != null) {
-			includedQueues = evalIncludeFilter(includeFilters.getType(), allQueues, includeFilters.getValues());
-		}
-		QueueExcludeFilters excludeFilters = this.queueManager.getQueueExcludeFilters();
-		if (excludeFilters != null) {
-			queueList = evalExcludeFilter(excludeFilters.getType(), includedQueues, excludeFilters.getValues());
-		} else {
-			queueList = includedQueues;
-		}
-		logger.debug("Queues after filter: " + Arrays.toString(queueList.toArray()));
-	}
-
-	@Override
 	protected void publishMetrics() throws TaskExecutionException {
-		if (queueList == null || queueList.isEmpty()) {
-			logger.debug("queue List empty");
-			return;
-		}
-		for (String queueName : queueList) {
-			publishQueueMetrics(queueName);
-		}
-	}
-
-	private void publishQueueMetrics(String queueName) {
 		/*
 		 * attrs = { CMQC.MQCA_Q_NAME, CMQC.MQIA_CURRENT_Q_DEPTH, CMQC.MQIA_MAX_Q_DEPTH, CMQC.MQIA_OPEN_INPUT_COUNT, CMQC.MQIA_OPEN_OUTPUT_COUNT };
 		 */
 		long entryTime = System.currentTimeMillis();
 
 		if (getMetricsToReport() == null || getMetricsToReport().isEmpty()) {
-			logger.debug("Queue metrics to report is null or empty, nothing to publish");
+			logger.debug("Queue metrics to report from the config is null or empty, nothing to publish");
 			return;
 		}
 
-		logger.debug("publishQueueMetrics entry time for queue {} is {} milliseconds", queueName, entryTime);
 
 		int[] attrs = getIntArrtibutesArray(CMQC.MQCA_Q_NAME);
+		logger.debug("Attributes being sent along PCF agent request to query channel metrics: " + Arrays.toString(attrs));
 
-		PCFMessage request = new PCFMessage(CMQCFC.MQCMD_INQUIRE_Q);
-		request.addParameter(CMQC.MQCA_Q_NAME, queueName);
-		request.addParameter(CMQC.MQIA_Q_TYPE, CMQC.MQQT_ALL);
-		request.addParameter(CMQCFC.MQIACF_Q_ATTRS, attrs);
-		PCFMessage[] response;
+		Set<String> queueGenericNames = this.queueManager.getQueueFilters().getInclude();
+		for(String queueGenericName : queueGenericNames){
+			// list of all metrics extracted through MQCMD_INQUIRE_Q is mentioned here https://www.ibm.com/support/knowledgecenter/SSFKSJ_7.5.0/com.ibm.mq.ref.adm.doc/q087810_.htm
+			PCFMessage request = new PCFMessage(CMQCFC.MQCMD_INQUIRE_Q);
+			request.addParameter(CMQC.MQCA_Q_NAME, queueGenericName);
+			request.addParameter(CMQC.MQIA_Q_TYPE, CMQC.MQQT_ALL);
+			request.addParameter(CMQCFC.MQIACF_Q_ATTRS, attrs);
+			PCFMessage[] response;
 
-		try {
-			logger.debug("sending PCF agent request to query queue metrics");
-			long startTime = System.currentTimeMillis();
-			response = agent.send(request);
-			long endTime = System.currentTimeMillis() - startTime;
-			logger.debug("PCF agent queue metrics query response for {} received in {} milliseconds", queueName, endTime);
-			if (response == null || response.length <= 0) {
-				logger.debug("Unexpected Error while PCFMessage.send(), response is either null or empty");
-				return;
-			}
-			for (int i = 0; i < response.length; i++) {
-				Iterator<String> itr = getMetricsToReport().keySet().iterator();
-				while (itr.hasNext()) {
-					String metrickey = itr.next();
-					WMQMetricOverride wmqOverride = (WMQMetricOverride) getMetricsToReport().get(metrickey);
-					int metricVal = response[i].getIntParameterValue(wmqOverride.getConstantValue());
-					publishMetric(wmqOverride, metricVal, queueManager.getName(), getAtrifact(), queueName, wmqOverride.getAlias());
+			try {
+				logger.debug("sending PCF agent request to query metrics for generic queue {}",queueGenericName);
+				long startTime = System.currentTimeMillis();
+				response = agent.send(request);
+				long endTime = System.currentTimeMillis() - startTime;
+				logger.debug("PCF agent queue metrics query response for generic queue {} received in {} milliseconds", queueGenericName, endTime);
+				if (response == null || response.length <= 0) {
+					logger.debug("Unexpected Error while PCFMessage.send(), response is either null or empty");
+					return;
 				}
+				for (int i = 0; i < response.length; i++) {
+					String queueName = response[i].getStringParameterValue(CMQC.MQCA_Q_NAME).trim();
+					Set<ExcludeFilters> excludeFilters = this.queueManager.getQueueFilters().getExclude();
+					if(!isExcluded(queueName,excludeFilters)) { //check for exclude filters
+						logger.debug("Pulling out metrics for queue name {}",queueName);
+						Iterator<String> itr = getMetricsToReport().keySet().iterator();
+						while (itr.hasNext()) {
+							String metrickey = itr.next();
+							WMQMetricOverride wmqOverride = (WMQMetricOverride) getMetricsToReport().get(metrickey);
+							int metricVal = response[i].getIntParameterValue(wmqOverride.getConstantValue());
+							publishMetric(wmqOverride, metricVal, queueManager.getName(), getAtrifact(), queueName, wmqOverride.getAlias());
+						}
+					}
+					else{
+						logger.debug("Queue name {} is excluded.",queueName);
+					}
+				}
+			} catch (PCFException pcfe) {
+				logger.error("PCFException caught while collecting metric for Queue: " + queueGenericName, pcfe);
+				PCFMessage[] msgs = (PCFMessage[]) pcfe.exceptionSource;
+				for (int i = 0; i < msgs.length; i++) {
+					logger.error(msgs[i].toString());
+				}
+				// Dont throw exception as it will stop queuemetric colloection
+			} catch (Exception mqe) {
+				logger.error("MQException caught", mqe);
+				// Dont throw exception as it will stop queuemetric colloection
 			}
-		} catch (PCFException pcfe) {
-			logger.error("PCFException caught while collecting metric for Queue: " + queueName, pcfe);
-			PCFMessage[] msgs = (PCFMessage[]) pcfe.exceptionSource;
-			for (int i = 0; i < msgs.length; i++) {
-				logger.error(msgs[i].toString());
-			}
-			// Dont throw exception as it will stop queuemetric colloection
-		} catch (Exception mqe) {
-			logger.error("MQException caught", mqe);
-			// Dont throw exception as it will stop queuemetric colloection
-		} finally {
-			long exitTime = System.currentTimeMillis() - entryTime;
-			logger.debug("Time taken to publish metrics for queue {} is {} milliseconds", queueName, exitTime);
+			// #TODO extract metrics like CMQC.MQIA_MSG_DEQ_COUNT, CMQC.MQIA_MSG_ENQ_COUNT  using the CMQCFC.MQCMD_RESET_Q_STATS through a new PCF request
 		}
+		long exitTime = System.currentTimeMillis() - entryTime;
+		logger.debug("Time taken to publish metrics for all queues is {} milliseconds", exitTime);
 	}
+
+
 
 	@Override
 	public String getAtrifact() {
@@ -137,63 +114,5 @@ public class QueueMetricsCollector extends MetricsCollector {
 		return this.metricsToReport;
 	}
 
-	private List<String> getQueueList() throws TaskExecutionException {
-
-		List<String> allQueueList = new ArrayList<String>();
-		List<String> localQueueList = getQueueListByType(MQConstants.MQQT_LOCAL);
-
-		allQueueList.addAll(localQueueList);
-
-		return allQueueList;
-
-	}
-
-	/**
-	 * @param queueType
-	 *            May be one of following types type MQConstants.MQQT_LOCAL, MQConstants.MQQT_ALIAS, MQConstants.MQQT_REMOTE
-	 * @return List of queues of specified type
-	 * @throws TaskExecutionException
-	 */
-	private List<String> getQueueListByType(int queueType) throws TaskExecutionException {
-		List<String> queueList = new ArrayList<String>();
-
-		try {
-
-			PCFMessage inquireNames = new PCFMessage(CMQCFC.MQCMD_INQUIRE_Q_NAMES);
-
-			inquireNames.addParameter(CMQC.MQCA_Q_NAME, "*");
-
-			// TODO see if this filters out the model queues
-			inquireNames.addParameter(CMQC.MQIA_Q_TYPE, queueType);
-
-			PCFMessage[] responseMsgs = agent.send(inquireNames);
-
-			if (responseMsgs == null || responseMsgs.length == 0) {
-				logger.error("Unable to get response from PCF");
-				throw new TaskExecutionException("Unable to get response from PCF");
-			}
-			String[] names = (String[]) responseMsgs[0].getParameterValue(CMQCFC.MQCACF_Q_NAMES);
-
-			for (int i = 0; i < names.length; i++) {
-				names[i] = names[i].trim();
-
-				if (!queueList.contains(names[i])) {
-					queueList.add(names[i]);
-				}
-
-			}
-		} catch (PCFException e) {
-			logger.error(e.getMessage());
-			throw new TaskExecutionException(e);
-		} catch (MQException e) {
-			logger.error(e.getMessage());
-			throw new TaskExecutionException(e);
-		} catch (IOException e) {
-			logger.error(e.getMessage());
-			throw new TaskExecutionException(e);
-		}
-
-		return queueList;
-	}
 
 }
