+++ src/main/resources/conf/config.yaml
@@ -1,50 +1,55 @@
-# WebsphereMQ instance particulars
+### ANY CHANGES TO THIS FILE DOES NOT REQUIRE A RESTART ###
+numberOfThreads: 10
+
+#This will create this metric in all the tiers, under this path. Please make sure to have a trailing |
+#metricPrefix: Custom Metrics|WebsphereMQ|
+
+#This will create it in specific Tier. Replace <TIER_ID>. Please make sure to have a trailing |
+metricPrefix: "Server|Component:<TIER_ID>|Custom Metrics|WebsphereMQ|"
 
 queueManagers:
   - host: "localhost"
     port: 1414
     #Actual name of the queue manager
-    name: 
+    name: ""
     #Channel name of the queue manager, channel should be server-conn type.
     channelName: "SYSTEM.ADMIN.SVRCONN"
     #The transport type for the queue manager connection, the default is "Bindings" for a binding type connection
     #For bindings type connection WMQ extension (i.e machine agent) need to be on the same machine on which WebbsphereMQ server is running
     #for client type connection change it to "Client".
     transportType: "Bindings"
-    #user with admin level access, no need to provide credentials in case of bindings transport type, it is only applicable for client type
-    username:
-    password:
+    #user with admin level access, comment out the credentials in case of bindings transport type, it is only applicable for client type
+    username: ""
+    password: ""
     
     #SSL related properties
     # e.g."C:/Program Files (x86)/IBM/WebSphere MQ/ssl/key" , Please use forward slash in the path.
-    sslKeyRepository:  
+    sslKeyRepository: ""
     # e.g. "SSL_RSA_WITH_AES_128_CBC_SHA256"
-    cipherSuite:
-    
-    # First Include filter is applied and on the result of include filter exclude filter is applied.  
-    queueIncludeFilters:
-        #type value: STARTSWITH, EQUALS, ENDSWITH, CONTAINS, NONE (default all queues) 
-        type: "NONE"
-        #The name of the queue or queue name pattern as per queue filter, comma separated values
-        values: []
-          
-    queueExcludeFilters:
-        #type value: STARTSWITH, EQUALS, ENDSWITH, CONTAINS, NONE (default all queues) 
-        type: "STARTSWITH"
-        #The name of the queue or queue name pattern as per queue filter, comma separated values
-        values: ["SYSTEM", "AMQ", "KMQ"]
+    cipherSuite: ""
 
-    channelIncludeFilters:
-        #type value: STARTSWITH, EQUALS, ENDSWITH, CONTAINS, NONE (default all channels) 
-        type: "NONE"
-        #The name of the channel or channel name pattern as per channel filter, comma separated values
-        values: []    
+    queueFilters:
+        #Can provide complete queue name or generic names. A generic name is a character string followed by an asterisk (*),
+        #for example ABC*, and it selects all objects having names that start with the selected character string.
+        #An asterisk on its own matches all possible names.
+        include: ["*"]
+        exclude:
+             #type value: STARTSWITH, EQUALS, ENDSWITH, CONTAINS
+           - type: "STARTSWITH"
+             #The name of the queue or queue name pattern as per queue filter, comma separated values
+             values: ["SYSTEM","AMQ"]
 
-    channelExcludeFilters:
-        #type value: STARTSWITH, EQUALS, ENDSWITH, CONTAINS, NONE (default all channels) 
-        type: "STARTSWITH"
-        #The name of the channel or channel name pattern as per channel filter, comma separated values
-        values: ["SYSTEM", "AMQ", "KMQ"]
+
+    channelFilters:
+        #Can provide complete channel name or generic names. A generic name is a character string followed by an asterisk (*),
+        #for example ABC*, and it selects all objects having names that start with the selected character string.
+        #An asterisk on its own matches all possible names.
+        include: ["*"]
+        exclude:
+           #type value: STARTSWITH, EQUALS, ENDSWITH, CONTAINS
+           - type: "STARTSWITH"
+           #The name of the queue or queue name pattern as per queue filter, comma separated values
+             values: ["SYSTEM"]
     
 mqMertics:
   # This Object will extract queue manager metrics
@@ -78,7 +83,7 @@ mqMertics:
           ibmConstant: "com.ibm.mq.constants.CMQCFC.MQIACH_MSGS"
           
         - Status: "Status"
-          ibmConstant: "com.ibm.mq.constants.CMQCFC.MQIACH_CHANNEL_STATUS"
+          ibmConstant: "com.ibm.mq.constants.CMQCFC.MQIACH_CHANNEL_STATUS"  #http://www.ibm.com/support/knowledgecenter/SSFKSJ_7.5.0/com.ibm.mq.ref.dev.doc/q090880_.htm
           
         - ByteSent: "Byte Sent"
           ibmConstant: "com.ibm.mq.constants.CMQCFC.MQIACH_BYTES_SENT"
@@ -92,7 +97,4 @@ mqMertics:
         - BuffersReceived: "Buffers Received"
           ibmConstant: "com.ibm.mq.constants.CMQCFC.MQIACH_BUFFERS_RECEIVED"
   
-numberOfThreads: 10
-          
-#prefix used to show up metrics in AppDynamics, Note: it was root_category_name in monitor.xml
-metricPrefix: "Custom Metrics|WebsphereMQ|" 
+
