+++ src/main/resources/conf/monitor.xml
@@ -1,24 +1,24 @@
 <monitor>
-	<name>MqQueueMonitor</name>
+	<name>WMQueueMonitor</name>
 	<type>managed</type>
         <enabled>true</enabled>
-	<description>Monitors MQ Queues</description>
+	<description>Monitors Websphere Messaging Queues</description>
 	<monitor-configuration>
 	</monitor-configuration>
 	<monitor-run-task>
 		<execution-style>periodic</execution-style>
-		<execution-frequency-in-seconds>5</execution-frequency-in-seconds>
-		<name>MQ Queue Monitor Run Task</name>
-		<display-name>MQ Queue Monitor Task</display-name>
-		<description>MQ Queue Monitor Task</description>
+		<execution-frequency-in-seconds>30</execution-frequency-in-seconds>
+		<name>Websphere Messaging Queue Monitor Run Task</name>
+		<display-name>Websphere Messaging Queue Monitor Task</display-name>
+		<description>Websphere Messaging Queue Monitor Task</description>
 		<type>java</type>
 		<execution-timeout-in-secs>60</execution-timeout-in-secs>
 		<task-arguments>
-		    <argument name="config-file" is-required="false" default-value="monitors/MQMonitor/config.yaml"/>
+		    <argument name="config-file" is-required="false" default-value="monitors/WMQMonitor/config.yaml"/>
 		</task-arguments>
 		<java-task>
 			<classpath>websphere-mq-monitoring-extension.jar;com.ibm.mq.commonservices.jar;com.ibm.mq.jar;com.ibm.mq.jmqi.jar;dhbcore.jar;com.ibm.mq.headers.jar;connector.jar;com.ibm.mq.pcf.jar</classpath>
-			<impl-class>com.appdynamics.extensions.webspheremq.WebspherMqMonitor</impl-class>
+			<impl-class>com.appdynamics.extensions.webspheremq.WebsphereMQMonitor</impl-class>
 		</java-task>
 	</monitor-run-task>
 </monitor>
