+++ src/test/java/com/appdynamics/extensions/webspheremq/config/ConfigFileTest.java
@@ -2,7 +2,7 @@ package com.appdynamics.extensions.webspheremq.config;
 
 import org.junit.Test;
 
-import com.appdynamics.extensions.webspheremq.WMQauthenticationAndAuthorization;
+import com.appdynamics.extensions.webspheremq.WebsphereMQContext;
 import com.appdynamics.extensions.yml.YmlReader;
 
 public class ConfigFileTest {
@@ -16,14 +16,14 @@ public class ConfigFileTest {
 	public void loadClientNoHostConfigTest(){
 		Configuration config = YmlReader.readFromFile(this.getClass().getResource("/conf/configClientNoHost.yaml").getFile(),Configuration.class);
 		QueueManager queueManager = config.getQueueManagers()[0];
-		WMQauthenticationAndAuthorization auth = new WMQauthenticationAndAuthorization(queueManager);		
+		WebsphereMQContext auth = new WebsphereMQContext(queueManager);
 	}
 	
 	@Test(expected=IllegalArgumentException.class)
 	public void loadBindingNoQueuemanagerConfigTest(){
 		Configuration config = YmlReader.readFromFile(this.getClass().getResource("/conf/configBindingNoQm.yaml").getFile(),Configuration.class);
 		QueueManager queueManager = config.getQueueManagers()[0];
-		WMQauthenticationAndAuthorization auth = new WMQauthenticationAndAuthorization(queueManager);		
+		WebsphereMQContext auth = new WebsphereMQContext(queueManager);
 	}
