+++ src/test/resources/conf/configClientNoHost.yaml
@@ -1,10 +1,17 @@
-# WebsphereMQ instance particulars
+### ANY CHANGES TO THIS FILE DOES NOT REQUIRE A RESTART ###
+numberOfThreads: 10
+
+#This will create this metric in all the tiers, under this path. Please make sure to have a trailing |
+#metricPrefix: Custom Metrics|WebsphereMQ|
+
+#This will create it in specific Tier. Replace <TIER_NAME>. Please make sure to have a trailing |
+metricPrefix: "Server|Component:<TIER_NAME>|Custom Metrics|WebsphereMQ|"
 
 queueManagers:
   - host: ""
     
     #Actual name of the queue manager
-    name: 
+    name: ""
     #Channel name of the queue manager
     channelName: "SYSTEM.ADMIN.SVRCONN"
     #The transport type for the queue manager connection, the default is "Bindings" for a binding type connection
@@ -15,30 +22,27 @@ queueManagers:
     username:
     password:
     
-    # First Include filter is applied and on the result of include filter exclude filter is applied.  
-    queueIncludeFilters:
-        #type value: STARTSWITH, EQUALS, ENDSWITH, CONTAINS, NONE (default all queues) 
-        type: "NONE"
-        #The name of the queue or queue name pattern as per queue filter, comma separated values
-        values: ["SYSTEM", "AMQ", "KMQ"]
-          
-    queueExcludeFilters:
-        #type value: STARTSWITH, EQUALS, ENDSWITH, CONTAINS, NONE (default all queues) 
-        type: "STARTSWITH"
-        #The name of the queue or queue name pattern as per queue filter, comma separated values
-        values: ["SYSTEM", "AMQ", "KMQ"]
+    queueFilters:
+        #Can provide complete queue name or generic names. A generic name is a character string followed by an asterisk (*),
+        #for example ABC*, and it selects all objects having names that start with the selected character string.
+        #An asterisk on its own matches all possible names.
+        include: ["ABC","DEF*"]
+        exclude:
+            #type value: STARTSWITH, EQUALS, ENDSWITH, CONTAINS
+            - type: "STARTSWITH"
+            #The name of the queue or queue name pattern as per queue filter, comma separated values
+              values: ["SYSTEM","AMQ"]
 
-    channelIncludeFilters:
-        #type value: STARTSWITH, EQUALS, ENDSWITH, CONTAINS, NONE (default all queues) 
-        type: "NONE"
-        #The name of the channel or channel name pattern as per channel filter, comma separated values
-        values: ["SYSTEM", "AMQ", "KMQ"]    
-
-    channelExcludeFilters:
-        #type value: STARTSWITH, EQUALS, ENDSWITH, CONTAINS, NONE (default all queues) 
-        type: "STARTSWITH"
-        #The name of the channel or channel name pattern as per channel filter, comma separated values
-        values: ["SYSTEM", "AMQ", "KMQ"]
+    channelFilters:
+        #Can provide complete channel name or generic names. A generic name is a character string followed by an asterisk (*),
+        #for example ABC*, and it selects all objects having names that start with the selected character string.
+        #An asterisk on its own matches all possible names.
+        include: ["ABC","DEF*"]
+        exclude:
+            #type value: STARTSWITH, EQUALS, ENDSWITH, CONTAINS
+            - type: "STARTSWITH"
+            #The name of the queue or queue name pattern as per queue filter, comma separated values
+              values: ["SYSTEM","AMQ"]
     
 mqMertics:
   # This Object will extract queue manager metrics
@@ -72,7 +76,7 @@ mqMertics:
           ibmConstant: "com.ibm.mq.constants.CMQCFC.MQIACH_MSGS"
           
         - Status: "Status"
-          ibmConstant: "com.ibm.mq.constants.CMQCFC.MQIACH_CHANNEL_STATUS"
+          ibmConstant: "com.ibm.mq.constants.CMQCFC.MQIACH_CHANNEL_STATUS" #http://www.ibm.com/support/knowledgecenter/SSFKSJ_7.5.0/com.ibm.mq.ref.dev.doc/q090880_.htm
           
         - ByteSent: "Byte Sent"
           ibmConstant: "com.ibm.mq.constants.CMQCFC.MQIACH_BYTES_SENT"
@@ -86,7 +90,4 @@ mqMertics:
         - BuffersReceived: "Buffers Received"
           ibmConstant: "com.ibm.mq.constants.CMQCFC.MQIACH_BUFFERS_RECEIVED"
   
-numberOfThreads: 10
-          
-#prefix used to show up metrics in AppDynamics, Note: it was root_category_name in monitor.xml
-metricPrefix: "Custom Metrics|WebsphereMQ|" 
+
